Refactored code to use DI for controllers and for app
Ran into an issue with tests where failing do to accessing the same database state in different test suites which lead to a bunch of flaky tests. Found a solution to use transactions for each test which required using PoolClient which handles one client only, its impossible to do transactions with the regular Pool object because that uses multiple clients. This lead to me making the controllers take in the repo being using as an argument and for the app I made a buildApp function to help set up the app properly for the test environment. This allowed me to succesfully use PoolClients in the tests removing the flakiness.